fix EasyAdmin patch tests in CI
